Update README.md
Added User Registration and Login Page Using PHP and MySQL.
Remarks: The passwords that you enter are encrypted using password_hash() and PASSWORD_BCRYPT. It uses the CRYPT_BLOWFISH algorithm to create the hash. This will produce a standard crypt() compatible hash using the "$2y$" identifier. The result will always be a 60 character string, or false on failure.